🔧 Add `dotnetw` wrapper
Invokes `dotnet` in a Docker container with the project directory
mounted as a volume, with write permissions set on the output
directories, and with the working directory set to the mounted
project directory.

This makes it convenient to run `dotnet` commands in a local
development environment that closely resembles the CircleCI
build environment, making CI build outcomes more easily
predictable.